+++ R/col_prep.R
@@ -16,7 +16,8 @@ col_prep <- function(data, breaks_odds, breaks_stake, bound_odds, bound_stake) {
                   StakeMod   = dplyr::if_else(Stake > bound_stake, bound_stake, Stake),
                   StakeGroup = cut(StakeMod, breaks = breaks_stake, include.lowest = TRUE) %>% as.character(), 
                   ID         = dplyr::row_number() %>% rev(), 
-                  DaysDiff   = difftime(MatchDay, BetDay, units = "d") %>% as.numeric())
+                  DaysDiff   = difftime(MatchDay, BetDay, units = "d") %>% as.numeric(), 
+                  Bookmaker  = dplyr::if_else(is.na(Bookmaker), "Danske Spil", Bookmaker))
   
   df_count_bets <- data_tmp %>% 
     dplyr::mutate(ExposureBets = dplyr::if_else(is.na(Revenue) | is.na(Match), 0, 1)) %>%